+++ jlt/eigensystem.hpp
@@ -81,6 +81,69 @@ int symmetric_matrix_eigensystem(matrix<T>& A,
 
 
 template<class T>
+int hermitian_matrix_eigensystem(matrix<std::complex<T>>& A,
+				 std::vector<T>& eigvals)
+{
+  char jobz = 'V';	// 'N'-eigenvalues only, 'V'-eigenvalues and vectors
+  char uplo = 'L';	// 'L'ower or 'U'pper triangle stored (opposite)
+  int N = A.rows();	// Dimensions of matrix.
+
+  assert(N == static_cast<int>(A.columns()) && N == static_cast<int>(eigvals.size()));
+
+  int info;
+
+  // Use temporary vector and copy of A, since we need to reverse the order.
+  std::vector<T> eigs(N);
+  matrix<std::complex<T>> U(A);
+
+  // Real workspace for hermitian eigenvalues
+  int rworksize = std::max(1, 3*N-2);
+  std::vector<T> rwork(rworksize);
+
+#ifdef JLT_MIN_WORKSIZE
+  // Use the smallest possible workspace.
+  int cworksize = std::max(1, 2*N-1);
+  std::vector<std::complex<T>> cwork(cworksize);
+#else
+  // Call the routine with cworksize = -1, to get the ideal size of workspace.
+  int cworksize = -1;
+  std::complex<T> ctmpwork[1];
+
+  lapack::heev(&jobz, &uplo, &N, U.data(), &N, eigs.data(),
+	       ctmpwork, &cworksize, rwork.data(), &info);
+
+  // Now allocate the memory for the workspace.
+  cworksize = static_cast<int>(ctmpwork[0].real());
+
+#ifdef JLT_DEBUG
+  std::cerr << "jlt::hermitian_matrix_eigensystem:     cworksize = " << cworksize << std::endl;
+  std::cerr << "jlt::hermitian_matrix_eigensystem:     rworksize = " << rworksize << std::endl;
+  std::cerr << "jlt::hermitian_matrix_eigensystem: min cworksize = ";
+  std::cerr << std::max(1, 2*N-1) << std::endl;
+#endif
+
+  std::vector<std::complex<T>> cwork(cworksize);
+#endif
+
+  lapack::heev(&jobz, &uplo, &N, U.data(), &N, eigs.data(),
+	       cwork.data(), &cworksize, rwork.data(), &info);
+
+  // Output eigenvalues in *descending* order.
+  for (int i = 0; i < N; ++i) eigvals[i] = eigs[N-i-1];
+  // Also need to reverse the eigenvectors, stored as row vectors.
+  for (int i = 0; i < N; ++i)
+    {
+      for (int j = 0; j < N; ++j)
+	{
+	  A(i,j) = U(N-i-1,j);
+	}
+    }
+
+  return info;
+}
+
+
+template<class T>
 int matrix_eigenvalues(matrix<T>& A,
 		       std::vector<std::complex<T>>& eigvals)
 {